+++ src/app/configuration/home-textile/process-type/process-type.component.ts
@@ -5,6 +5,7 @@ import { ToastrService } from 'ngx-toastr';
 import { environment } from 'src/environments/environment';
 import { AddProcessTypeComponent } from './add-process-type/add-process-type.component';
 import { EditProcessTypeComponent } from './edit-process-type/edit-process-type.component';
+import Swal from 'sweetalert2/dist/sweetalert2.js';
 
 @Component({
   selector: 'app-process-type',
@@ -56,31 +57,69 @@ export class ProcessTypeComponent implements OnInit {
               }
             
             
-            
-              deleteProcess(id)
-              {
-                this.http.delete(`${environment.apiUrl}/api/TextileGarments/DeleteProcessType/`+id.id )
-                .subscribe(
-                  res=> { 
-                    this.response = res;
-                    if (this.response.success == true){
-                     this.toastr.error(this.response.message, 'Message.');
-                     this.fetch((data) => {
-                      this.rows = data;
-                    });
-                      
-                    }
-                    else {
-                      this.toastr.error('Something went Worng', 'Message.');
+  
+
+              deleteProcess(id){
+                Swal.fire({
+                  title: 'Are you sure?',
+                  text: "You won't be able to revert this!",
+                  icon: 'warning',
+                  showCancelButton: true,
+                  confirmButtonColor: '#3085d6',
+                  cancelButtonColor: '#d33',
+                  confirmButtonText: 'Delete it!'
+                }).then((result) => {
+                  if (result.isConfirmed) {
+                
+                    this.http.delete(`${environment.apiUrl}/api/TextileGarments/DeleteProcessType/`+id.id )
+                    .subscribe(
+                      res=> { 
+                        this.response = res;
+                        if (this.response.success == true){
+                         this.toastr.error(this.response.message, 'Message.');
+                         this.fetch((data) => {
+                          this.rows = data;
+                        });
+                          
                         }
-             
-                  }, err => {
-                    if (err.status == 400) {
-                      this.toastr.error(this.response.message, 'Message.');
-                    }
-                  });
-              }
+                        else {
+                          this.toastr.error('Something went Worng', 'Message.');
+                            }
+                 
+                      }, err => {
+                        if (err.status == 400) {
+                          this.toastr.error(this.response.message, 'Message.');
+                        }
+                      });
+                
+                    // Swal.fire(
+                    //   'Record',
+                    //   'Deleted Successfully.',
+                    //   'success'
+                    // )
+                  }
+                })
+                
+                }
             
+            
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
             
             
               addProcessTypeForm(){
