+++ src/app/configuration/product/payment-term/payment-term.component.ts
@@ -5,6 +5,7 @@ import { environment } from 'src/environments/environment';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { AddPaymentComponent } from './add-payment/add-payment.component';
 import { EditPaymentComponent } from './edit-payment/edit-payment.component';
+import Swal from 'sweetalert2/dist/sweetalert2.js';
 @Component({
   selector: 'app-payment-term',
   templateUrl: './payment-term.component.html',
@@ -57,31 +58,58 @@ export class PaymentTermComponent implements OnInit {
   
   
   
-    deletePayment(id)
-    {
-      this.http.delete(`${environment.apiUrl}/api/Products/DeletePaymentTerm/`+id.id )
-      .subscribe(
-        res=> { 
-          this.response = res;
-          if (this.response.success == true){
-           this.toastr.error(this.response.message, 'Message.');
-           this.fetch((data) => {
-            this.rows = data;
-          });
-            
-          }
-          else {
-            this.toastr.error('Something went Worng', 'Message.');
+
+
+    deletePayment(id){
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+      
+          this.http.delete(`${environment.apiUrl}/api/Products/DeletePaymentTerm/`+id.id )
+          .subscribe(
+            res=> { 
+              this.response = res;
+              if (this.response.success == true){
+               this.toastr.error(this.response.message, 'Message.');
+               this.fetch((data) => {
+                this.rows = data;
+              });
+                
               }
-   
-        }, err => {
-          if (err.status == 400) {
-            this.toastr.error(this.response.message, 'Message.');
-          }
-        });
-    }
+              else {
+                this.toastr.error('Something went Worng', 'Message.');
+                  }
+       
+            }, err => {
+              if (err.status == 400) {
+                this.toastr.error(this.response.message, 'Message.');
+              }
+            });
+      
+          // Swal.fire(
+          //   'Record',
+          //   'Deleted Successfully.',
+          //   'success'
+          // )
+        }
+      })
+      
+      }
   
   
+  
+
+
+
+
+
   
     addPaymentForm(){
       const modalRef = this.modalService.open(AddPaymentComponent, { centered: true });
